# backend/internal/handler/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger specification for a Duels back-end.",
        "title": "Duels API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/user/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's profile data based on their JWT token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user data",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user": {
                                    "$ref": "#/definitions/duels-api_internal_model.User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/user/profile-picture": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an authenticated user to upload a new profile image.\nThe uploaded image will replace the existing profile picture and return its new public URL.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Set user profile picture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Profile image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile picture successfully updated",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "image_url": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data or missing image file",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during image upload",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/user/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aggregated duel stats for the authenticated user (participated, wins/losses, earned/lost, etc.)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user duel stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user duel stats",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "stats": {
                                    "$ref": "#/definitions/duels-api_internal_model.UserStats"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/user/upload-images": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an authenticated user to upload one or more image files.\nEach image is stored, and the service returns an array of accessible URLs.\nUpload rate is limited via middleware to prevent spam (max 3 per minute).",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Upload user images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "Array of images to upload",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Images uploaded successfully, returns list of image URLs",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "image_urls": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data or missing image files",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "429": {
                        "description": "Too many uploads - rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during file save",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/user/username": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the username for the currently authenticated user. The new username must be valid and not already taken.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Change the username of the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Username change payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/duels-api_internal_model.UsernameChange"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Username changed successfully"
                    },
                    "400": {
                        "description": "Invalid request parameters or invalid username",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Authentication required or invalid token",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "409": {
                        "description": "This username is already taken",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during username change",
                        "schema": {
                            "$ref": "#/definitions/duels-api_pkg_apperrors.ErrorPublic"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "duels-api_internal_model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "public_address": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "duels-api_internal_model.UserStats": {
            "type": "object",
            "properties": {
                "earned_amount": {
                    "type": "number"
                },
                "losses_count": {
                    "type": "integer"
                },
                "lost_amount": {
                    "type": "number"
                },
                "net_profit": {
                    "type": "number"
                },
                "participated": {
                    "type": "integer"
                },
                "refunded_amount": {
                    "type": "number"
                },
                "refunded_count": {
                    "type": "integer"
                },
                "wins_count": {
                    "type": "integer"
                }
            }
        },
        "duels-api_internal_model.UsernameChange": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
        "duels-api_pkg_apperrors.ErrorPublic": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}